Refactored observation loaders
The `observations` submodule of `io` was almost untouched since I
started tastro from the previous library. The goal of this refactor was
to give the submodule some structure, and to make it easier to extend.

Each observable type now has a subdirectory with utilities to load
observations from source files, and pack them into a data structure. All
the data structures derive from a base class in `common.py`.

I moved the functionality to load IFMS and ODF data to the IFMSLoader
and ODFLoader classes, but the code to load the raw observations from
the source files is basically the same as before. The loader for Fdets
is just a prototype, and it is still important functionality, such as
being able to adjust the base frequency or being more robust in terms of
the structure of the files it can load. I will work on these updates the
moment I get a first iteration of the open-loop estimation.

Some minor updates:

- `observations_per_station` is now a dictionary of dictionaries that
  are extended with the `dict.update` method (doppler loaders)
- ODFLoader now includes an `observable_type` argument to be compatible
  with both regular and IFMS-like ODF files (sign adjustment)