Open-loop compatibility: config updates
I updated the configuration file to be compatible with open-loop
observations. This commit includes updates to the `config` submodule to
be able to work with the new configuration options.

Similar to other submodules, I created individual submodules in
`config` for the different observable types, and deprecated the scripts
they replace.

I moved the definition of settings for uplink frequency calculators to
the environment, but the code is still compatible with old configuration
files in which they are defined in the estimation section. I did this to
avoid having to repeat the uplink configuration for open-loop and
closed-loop doppler.

I updated the filter configuration with an optional `station` argument,
that is set to `all` by default. When the filter is not set to `all`, it
is only applied to observations performed from the chosen station. This
is needed to be able to filter open-loop observations between epochs, as
observations from multiple antennas are overlaping. I also updated the
uplink frequency setup to be able to specify ramp rates, which were
originally hard-coded to be equal to zero.

Unrelated to the new observation type, I also updated the base class for
the data structures such that containers (lists, dictionaries) are not
set to None when the associated part of the configuration file is
missing. This lets me have empty lists of sources, filters... which was
becoming problematic now that calling an unset argument results into a
fatal error.