Settings generator for estimator and environment
I refactored the functions to generate and update the system of bodies,
so that all the configuration is done regardless of the propagation or
the estimation being performed. I isolated the code to update the
environment into functions, and updated the part that defines settings
for the vehicle systems to also include the uplink frequency
information, which is now specified in the environment. I also added
some more debug information about the generation of the environment.

In the estimation module, I turned `ObservationModelSettingsGenerator`
into an abstract class, updated the function to generate link ends to
(optionally) avoid having to specified the receiver reference point when
it is the same as the id of the link, and I did some refactoring in the
`EstimationManager` to remove duplicated code and make it easier to
introduce new observable types.

The new `OpenLoopSettingsGenerator` is a clone of the closed-loop
equivalent. It would be a good idea to make these inherit from a base
class to avoid duplicating functionality, but the code I have is good
enough for now.

Important to note that the current setup results into an error when
trying to run the estimation, so it will have to be modified in coming
commits.